Renamed to downstream-http-path, and included flag for downstream-http-replace-path
Additional flag was required to determine if we should use the URI as part of the resolved URI (including path). This flag is required because for Infura use-case the / appended the the project id causes the authentication to fail. Test cases expanded to support:
path and replace - e.g. /v3/projectid would be passed to the external server
path and not replace - e.g. /v3/projectid would be passed along with login (path becomes /v3/projectid/login)
no path and not replace - this is the default functionality without arguments and is the current ethsigner fucntionality